Add AddFace method
This adds a method AddFace() to the multiface.Face that can be used to
add arbitrary (well-behaved) font.Face to the multiface.Face.

I initially omitted this function to make sure no one accidentally uses
it with Truetype font instead of AddTruetypeFace(). (Those fonts return
ok = true even when glyphs are missing, so they need to be handled
differently.)

Now #1 requests this to be usable with BDF fonts, so I added that
function and also a test with such a font.